Bug 1669181 - Part 66: Split uint8_clamped from vm/ArrayBufferObject.h. r=jandem
