+++ js/src/jit/MIR.cpp
@@ -32,6 +32,7 @@
 #include "util/Text.h"
 #include "util/Unicode.h"
 #include "vm/PlainObject.h"  // js::PlainObject
+#include "vm/Uint8Clamped.h"
 #include "wasm/WasmCode.h"
 
 #include "builtin/Boolean-inl.h"
